// UI/src/app/models/login.ts
export class logindata {
  email: string;
  password: string;
}

export class forgotdata {
  email: string;
  base_url: string;
}
export class objectives {
  objective_name: string;
  objective_description : string;
  objective_id: string;
}
 
export class deleteobjective {
  objective_id: string;
}